fix: Improve Exception messages
The previous code used `str(exc)` which does not display the Exception
type. The new message is much more informative. It also now includes
traceback information.

One additional test added.

Closes #275
