+++ httpstan/views.py
@@ -11,6 +11,7 @@ import json
 import logging
 import re
 import sqlite3
+import traceback
 from typing import Optional, Sequence, cast
 
 import aiohttp.web
@@ -285,7 +286,10 @@ async def handle_create_fit(request: aiohttp.web.Request) -> aiohttp.web.Respons
         if exc:
             # e.g., "hmc_nuts_diag_e_adapt_wrapper() got an unexpected keyword argument, ..."
             # e.g., "Found negative dimension size in variable declaration"
-            message, status = f"Error calling services function: `{exc}`", 400
+            message, status = (
+                f"Exception during call to services function: `{repr(exc)}`, traceback: `{traceback.format_tb(exc.__traceback__)}`",
+                400,
+            )
             logger.critical(message)
             operation["result"] = _make_error(message, status=status)
         else:
